+++ client/src/components/categorical/util.js
@@ -1,5 +1,5 @@
 // jshint esversion: 6
-export const alphabeticallySortedValues = values =>
+export default values =>
   Object.keys(values).sort((a, b) => {
     const textA = a.toUpperCase();
     const textB = b.toUpperCase();
