+++ client/src/components/continuous/continuous.js
@@ -5,18 +5,16 @@ import React from "react";
 import _ from "lodash";
 import { connect } from "react-redux";
 
-import HistogramBrush from "../brushableHistogram/";
+import HistogramBrush from "../brushableHistogram";
 
-@connect(state => {
-  return {
-    ranges: _.get(state.controls.world, "summary.obs", null),
-    metadata: _.get(state.controls.world, "obsAnnotations", null),
-    colorAccessor: state.controls.colorAccessor,
-    colorScale: state.controls.colorScale,
-    selectionUpdate: _.get(state.controls, "crossfilter.updateTime", null),
-    schema: _.get(state.controls.world, "schema", null)
-  };
-})
+@connect(state => ({
+  ranges: _.get(state.controls.world, "summary.obs", null),
+  metadata: _.get(state.controls.world, "obsAnnotations", null),
+  colorAccessor: state.controls.colorAccessor,
+  colorScale: state.controls.colorScale,
+  selectionUpdate: _.get(state.controls, "crossfilter.updateTime", null),
+  schema: _.get(state.controls.world, "schema", null)
+}))
 class Continuous extends React.Component {
   constructor(props) {
     super(props);
@@ -25,6 +23,8 @@ class Continuous extends React.Component {
 
     this.state = {};
   }
+
+  componentDidUpdate() {}
 
   handleColorAction(key) {
     return () => {
@@ -40,9 +40,10 @@ class Continuous extends React.Component {
   render() {
     const { ranges, obsAnnotations, schema } = this.props;
     if (schema && !this.continuousChecked) {
-      this.hasContinuous = _.some(schema.annotations.obs, d => {
-        return d.type === "int32" || d.type === "float32";
-      });
+      this.hasContinuous = _.some(
+        schema.annotations.obs,
+        d => d.type === "int32" || d.type === "float32"
+      );
       this.continuousChecked = true; /* only do this once */
     }
