+++ client/src/util/typedCrossfilter/index.js
@@ -279,14 +279,14 @@ class ScalarDimension {
     let found = 0;
 
     // skip up to offset records
-    for (i = len - 1; 0 <= i && skip < offset; i -= 1) {
+    for (i = len - 1; i >= 0 && skip < offset; i -= 1) {
       if (selection.isSelected(index[i])) {
         skip += 1;
       }
     }
 
     // grab up to k records
-    for (; 0 <= i && found < k; i -= 1) {
+    for (; i >= 0 && found < k; i -= 1) {
       if (selection.isSelected(index[i])) {
         ret.push(data[index[i]]);
         found += 1;
@@ -539,7 +539,7 @@ class ScalarGroup {
 
   // set the reduce functions to count records.
   reduceCount() {
-    return this.reduce((p, v) => p + 1, (p, v) => p - 1, () => 0);
+    return this.reduce(p => p + 1, p => p - 1, () => 0);
   }
 
   // set the reduce functions to sum records using specified value accessor.
